Add gene plot to an LDheatmap object.

Retrieve genes from the UCSC Genome Browser, generate the genes plot and add it to an LDheatmap object.

LDheatmap.addGenes(LDheatmap, chromosome, genome = NULL, genesLocation = 0.02,
splice_variants = TRUE, non_coding = TRUE)

Arguments

chromosome
A character string that identifies the chromosome.

genome
The genome assembly to use. The default is the most recent human genome assembly on the UCSC genome browser.

genesLocation
The gene plot distance from the LD heat map gene map.

splice_variants
If FALSE, exclude gene splice variants.

non_coding
If FALSE, exclude non-coding genes.

LDHeatmap
An object of class LDheatmap.

Value

An object of class LDheatmap given as an argument, with the grob LDheatmapGrob modified to inclue the "transcripts" child grob.

Details

Note: The LDheatmap object should have a non-NULL genetic.distances component. Otherwise the gene map will not be placed correctly. The genes are color coded as follows: black -- feature has a corresponding entry in the Protein Data Bank (PDB); dark blue -- transcript has been reviewed or validated by either the RefSeq, SwissProt or CCDS staff; medium blue -- other RefSeq transcripts; and light blue -- non-RefSeq transcripts.

For assemblies older than hg18, all genes are plotted in grey.

References

http://genome.ucsc.edu/cgi-bin/hgTrackUi?g=knownGene

See also

Examples

# NOT RUN {
data(GIMAP5.CEU)
ll<-LDheatmap(GIMAP5.CEU$snp.data,GIMAP5.CEU$snp.support$Position,flip=TRUE)
# Add gene plot
llplusgenes <- LDheatmap.addGenes(ll, chr="chr7", genome="hg18")
# }